Add WebAssembly (Emscripten) build target
ZathuraDbg now compiles to wasm and runs fully client-side in the browser.
Emulation-only: assembling (Keystone), emulating (icicle interpreter),
stepping, breakpoints, registers, memory and stack all work.

Build system:
- if(EMSCRIPTEN) branches in src/CMakeLists.txt link the prebuilt wasm
  libicicle.a / libkeystone.a / libcapstone.a (paths via -DWASM_LIBS_ROOT),
  embed the staged MEMFS (fonts + sleigh specs), and emit a fullscreen
  Zathura.html from wasm-shell.html.
- src/build-wasm.sh wraps emcmake configure + build.

App changes (all gated behind __EMSCRIPTEN__, native path unchanged):
- main.cpp: emscripten_set_main_loop instead of the while loop; WebGL2/GLES3
  GLFW hints; fixed MEMFS paths instead of whereami; GHIDRA_SRC=/ghidra;
  ImGui_ImplGlfw_InstallEmscriptenCallbacks for canvas-tracking fullscreen.
- app.cpp: first-run DockBuilder default layout so the UI fills the viewport
  with no saved imgui.ini.
- actions.cpp: synchronous executeInBackground (single-threaded); browser
  openBrowser/update-check; renamed the `times` global (clashed with libc).
- interpreter.hpp/execution.cpp/actions.cpp: waitForDebugReady() helper that
  no-ops in the single-threaded wasm build (the CV waits would deadlock).
- gdbRemoteStub.cpp / tinyfd_stub.c: inert replacements for the socket- and
  dialog-dependent sources, swapped in by CMake for the wasm build.

Portability fixes (apply to native too): fixed-size array instead of a VLA
with initializer, uint64->size_t cast, and a plain loop instead of
std::views::keys.

See WASM.md for build and hosting instructions.

+++ src/app/windows/memoryMapView.cpp
@@ -223,7 +223,7 @@ void memoryMapWindow()
                 if (ImGui::TextLink(ss.str().c_str()))
                 {
                     memoryEditorWindow.OptShowAddWindowButton = false;
-                    newMemEditWindowsInfo memWindowInfo = {memoryEditorWindow, memInfo[i].address, memInfo[i].size};
+                    newMemEditWindowsInfo memWindowInfo = {memoryEditorWindow, memInfo[i].address, static_cast<size_t>(memInfo[i].size)};
                     newMemEditWindows.push_back(memWindowInfo);
                 }
                 ImGui::PopStyleColor();
